Rename komponent OppsummeringVedtak til Vedtak
